franka_test/requirements.txt
-f https://developer.intel.com/ipex-whl-stable-xpu
matplotlib==3.4.3
numpy #==1.21.4
numba
opencv-python-headless #==4.5.5
psutil #==5.8.0
pybullet #==3.1.7
PyYAML #==6.0
scikit-learn #==1.2.0
scipy #==1.7.3
termcolor #==2.2.0
torch==2.0.1
faster_fifo
pyqtgraph
pyqt5